fix: add type annotations for PHPStan level 6 (batch 2 - 44 more files)
Continues from prior commit. Fixes missingType.* errors across API routes,
Helper classes, Map components, Messages, Migration, Model, Repository,
Service, View, and CustomPostType classes. Error count reduced from 201 to 74.

+++ src/Model/Calendar.php
@@ -102,8 +102,8 @@ class Calendar {
 		} else {
 			$weeks = array();
 			while ( $startDate <= $endDate ) {
-				$dayOfYear = date( 'z', $startDate );
-				$year      = date( 'Y', $startDate );
+				$dayOfYear = (int) date( 'z', $startDate );
+				$year      = (int) date( 'Y', $startDate );
 				$weeks[]   = new Week(
 					$year,
 					$dayOfYear,
